Refactor models.py
